refactor(ci): colocate vLLM interface analyzer
Move the analyzer package and its source-only unit tests under tests/e2e/vllm_interface so the complete upstream compatibility check is reviewed and executed from one directory tree. Update module paths, CLI invocation, and documentation without changing analysis semantics.
